[#700] Make CBP state machine work
