+++ Sources/ZcashLightClientKit/Block/Actions/Action.swift
@@ -57,7 +57,7 @@ protocol Action {
     // Each action updates context accordingly. It should at least set new state. Reason for this is that action can return different states for
     // different conditions. And action is the thing that knows these conditions.
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext
 
     // Should be called on each existing action when processor wants to stop. Some actions may do it's own background work.
     func stop() async
