+++ Sources/ZcashLightClientKit/Block/Actions/ChecksBeforeSyncAction.swift
@@ -44,7 +44,7 @@ class ChecksBeforeSyncAction {
 extension ChecksBeforeSyncAction: Action {
     var removeBlocksCacheWhenFailed: Bool { false }
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext {
         // clear any present cached state if needed.
         // this checks if there was a sync in progress that was
         // interrupted abruptly and cache was not able to be cleared
