+++ Sources/ZcashLightClientKit/Block/Actions/ClearAlreadyScannedBlocksAction.swift
@@ -19,7 +19,7 @@ class ClearAlreadyScannedBlocksAction {
 extension ClearAlreadyScannedBlocksAction: Action {
     var removeBlocksCacheWhenFailed: Bool { false }
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext {
         let lastScannedHeight = try await transactionRepository.lastScannedHeight()
         try await storage.clear(upTo: lastScannedHeight)
