+++ Sources/ZcashLightClientKit/Block/Actions/ClearCacheAction.swift
@@ -17,7 +17,7 @@ class ClearCacheAction {
 extension ClearCacheAction: Action {
     var removeBlocksCacheWhenFailed: Bool { false }
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext {
         try await storage.clear()
         await context.update(state: .finished)
         return context
