+++ Sources/ZcashLightClientKit/Block/Actions/ScanAction.swift
@@ -8,12 +8,12 @@
 import Foundation
 
 class ScanAction {
-    let config: CompactBlockProcessorNG.Configuration
+    let config: CompactBlockProcessor.Configuration
     let blockScanner: BlockScanner
     let logger: Logger
     let transactionRepository: TransactionRepository
 
-    init(container: DIContainer, config: CompactBlockProcessorNG.Configuration) {
+    init(container: DIContainer, config: CompactBlockProcessor.Configuration) {
         self.config = config
         blockScanner = container.resolve(BlockScanner.self)
         transactionRepository = container.resolve(TransactionRepository.self)
@@ -29,7 +29,7 @@ class ScanAction {
 extension ScanAction: Action {
     var removeBlocksCacheWhenFailed: Bool { true }
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext {
         guard let scanRange = await context.syncRanges.downloadAndScanRange else {
             return await update(context: context)
         }
@@ -38,6 +38,12 @@ extension ScanAction: Action {
         // This action is executed for each batch (batch size is 100 blocks by default) until all the blocks in whole `scanRange` are scanned.
         // So the right range for this batch must be computed.
         let batchRangeStart = max(scanRange.lowerBound, lastScannedHeight)
+        let batchRangeEnd = min(scanRange.upperBound, batchRangeStart + config.batchSize)
+
+        guard batchRangeStart <= batchRangeEnd else {
+            return await update(context: context)
+        }
+
         let batchRange = batchRangeStart...batchRangeStart + config.batchSize
         
         logger.debug("Starting scan blocks with range: \(batchRange.lowerBound)...\(batchRange.upperBound)")
@@ -51,8 +57,8 @@ extension ScanAction: Action {
             self?.logger.debug("progress: \(progress)")
             await didUpdate(.progressUpdated(.syncing(progress)))
         }
-        await context.update(state: .clearAlreadyScannedBlocks)
-        return context
+
+        return await update(context: context)
     }
 
     func stop() async { }
