+++ Sources/ZcashLightClientKit/Block/Actions/ScanDownloadedButUnscannedAction.swift
@@ -20,7 +20,7 @@ class ScanDownloadedButUnscannedAction {
 extension ScanDownloadedButUnscannedAction: Action {
     var removeBlocksCacheWhenFailed: Bool { false }
 
-    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessorNG.Event) async -> Void) async throws -> ActionContext {
+    func run(with context: ActionContext, didUpdate: @escaping (CompactBlockProcessor.Event) async -> Void) async throws -> ActionContext {
         if let range = await context.syncRanges.downloadedButUnscannedRange {
             logger.debug("Starting scan with downloaded but not scanned blocks with range: \(range.lowerBound)...\(range.upperBound)")
             let totalProgressRange = await context.totalProgressRange
