--- Sources/ZcashLightClientKit/Block/CompactBlockProcessorNG.swift
@@ -1,743 +1,0 @@
-//
-//  CompactBlockProcessor.swift
-//  
-//
-//  Created by Michal Fousek on 05.05.2023.
-//
-
-import Foundation
-
-actor CompactBlockProcessorNG {
-    // It would be better to use Combine here but Combine doesn't work great with async. When this runs regularly only one closure is stored here
-    // and that is one provided by `SDKSynchronizer`. But while running tests more "subscribers" is required here. Therefore it's required to handle
-    // more closures here.
-    private var eventClosures: [String: EventClosure] = [:]
-
-    private var syncTask: Task<Void, Error>?
-
-    private let actions: [CBPState: Action]
-    private var context: ActionContext
-
-    private(set) var config: Configuration
-    private var afterSyncHooksManager = AfterSyncHooksManager()
-
-    private let accountRepository: AccountRepository
-    private let blockDownloaderService: BlockDownloaderService
-    private let internalSyncProgress: InternalSyncProgress
-    private let latestBlocksDataProvider: LatestBlocksDataProvider
-    private let logger: Logger
-    private let metrics: SDKMetrics
-    private let rustBackend: ZcashRustBackendWelding
-    private let service: LightWalletService
-    private let storage: CompactBlockRepository
-    private let transactionRepository: TransactionRepository
-
-    private var retryAttempts: Int = 0
-    private var backoffTimer: Timer?
-    private var consecutiveChainValidationErrors: Int = 0
-
-    /// Compact Block Processor configuration
-    ///
-    /// - parameter fsBlockCacheRoot: absolute root path where the filesystem block cache will be stored.
-    /// - parameter dataDb: absolute file path of the DB where all information derived from the cache DB is stored.
-    /// - parameter spendParamsURL: absolute file path of the sapling-spend.params file
-    /// - parameter outputParamsURL: absolute file path of the sapling-output.params file
-    struct Configuration {
-        let alias: ZcashSynchronizerAlias
-        let saplingParamsSourceURL: SaplingParamsSourceURL
-        let fsBlockCacheRoot: URL
-        let dataDb: URL
-        let spendParamsURL: URL
-        let outputParamsURL: URL
-        let enhanceBatchSize: Int
-        let batchSize: Int
-        let retries: Int
-        let maxBackoffInterval: TimeInterval
-        let maxReorgSize = ZcashSDK.maxReorgSize
-        let rewindDistance: Int
-        let walletBirthdayProvider: () -> BlockHeight
-        var walletBirthday: BlockHeight { walletBirthdayProvider() }
-        let downloadBufferSize: Int = 10
-        let network: ZcashNetwork
-        let saplingActivation: BlockHeight
-        let cacheDbURL: URL?
-        var blockPollInterval: TimeInterval {
-            TimeInterval.random(in: ZcashSDK.defaultPollInterval / 2 ... ZcashSDK.defaultPollInterval * 1.5)
-        }
-
-        init(
-            alias: ZcashSynchronizerAlias,
-            cacheDbURL: URL? = nil,
-            fsBlockCacheRoot: URL,
-            dataDb: URL,
-            spendParamsURL: URL,
-            outputParamsURL: URL,
-            saplingParamsSourceURL: SaplingParamsSourceURL,
-            enhanceBatchSize: Int = ZcashSDK.DefaultEnhanceBatch,
-            batchSize: Int = ZcashSDK.DefaultSyncBatch,
-            retries: Int = ZcashSDK.defaultRetries,
-            maxBackoffInterval: TimeInterval = ZcashSDK.defaultMaxBackOffInterval,
-            rewindDistance: Int = ZcashSDK.defaultRewindDistance,
-            walletBirthdayProvider: @escaping () -> BlockHeight,
-            saplingActivation: BlockHeight,
-            network: ZcashNetwork
-        ) {
-            self.alias = alias
-            self.fsBlockCacheRoot = fsBlockCacheRoot
-            self.dataDb = dataDb
-            self.spendParamsURL = spendParamsURL
-            self.outputParamsURL = outputParamsURL
-            self.saplingParamsSourceURL = saplingParamsSourceURL
-            self.network = network
-            self.enhanceBatchSize = enhanceBatchSize
-            self.batchSize = batchSize
-            self.retries = retries
-            self.maxBackoffInterval = maxBackoffInterval
-            self.rewindDistance = rewindDistance
-            self.walletBirthdayProvider = walletBirthdayProvider
-            self.saplingActivation = saplingActivation
-            self.cacheDbURL = cacheDbURL
-        }
-
-        init(
-            alias: ZcashSynchronizerAlias,
-            fsBlockCacheRoot: URL,
-            dataDb: URL,
-            spendParamsURL: URL,
-            outputParamsURL: URL,
-            saplingParamsSourceURL: SaplingParamsSourceURL,
-            enhanceBatchSize: Int = ZcashSDK.DefaultEnhanceBatch,
-            batchSize: Int = ZcashSDK.DefaultSyncBatch,
-            retries: Int = ZcashSDK.defaultRetries,
-            maxBackoffInterval: TimeInterval = ZcashSDK.defaultMaxBackOffInterval,
-            rewindDistance: Int = ZcashSDK.defaultRewindDistance,
-            walletBirthdayProvider: @escaping () -> BlockHeight,
-            network: ZcashNetwork
-        ) {
-            self.alias = alias
-            self.fsBlockCacheRoot = fsBlockCacheRoot
-            self.dataDb = dataDb
-            self.spendParamsURL = spendParamsURL
-            self.outputParamsURL = outputParamsURL
-            self.saplingParamsSourceURL = saplingParamsSourceURL
-            self.walletBirthdayProvider = walletBirthdayProvider
-            self.saplingActivation = network.constants.saplingActivationHeight
-            self.network = network
-            self.cacheDbURL = nil
-            self.enhanceBatchSize = enhanceBatchSize
-            self.batchSize = batchSize
-            self.retries = retries
-            self.maxBackoffInterval = maxBackoffInterval
-            self.rewindDistance = rewindDistance
-        }
-    }
-
-    /// Initializes a CompactBlockProcessor instance
-    /// - Parameters:
-    ///  - service: concrete implementation of `LightWalletService` protocol
-    ///  - storage: concrete implementation of `CompactBlockRepository` protocol
-    ///  - backend: a class that complies to `ZcashRustBackendWelding`
-    ///  - config: `Configuration` struct for this processor
-    init(container: DIContainer, config: Configuration) {
-        self.init(
-            container: container,
-            config: config,
-            accountRepository: AccountRepositoryBuilder.build(dataDbURL: config.dataDb, readOnly: true, logger: container.resolve(Logger.self))
-        )
-    }
-
-    /// Initializes a CompactBlockProcessor instance from an Initialized object
-    /// - Parameters:
-    ///     - initializer: an instance that complies to CompactBlockDownloading protocol
-    init(initializer: Initializer, walletBirthdayProvider: @escaping () -> BlockHeight) {
-        self.init(
-            container: initializer.container,
-            config: Configuration(
-                alias: initializer.alias,
-                fsBlockCacheRoot: initializer.fsBlockDbRoot,
-                dataDb: initializer.dataDbURL,
-                spendParamsURL: initializer.spendParamsURL,
-                outputParamsURL: initializer.outputParamsURL,
-                saplingParamsSourceURL: initializer.saplingParamsSourceURL,
-                walletBirthdayProvider: walletBirthdayProvider,
-                network: initializer.network
-            ),
-            accountRepository: initializer.accountRepository
-        )
-    }
-
-    init(container: DIContainer, config: Configuration, accountRepository: AccountRepository) {
-//        Dependencies.setupCompactBlockProcessor(
-//            in: container,
-//            config: config,
-//            accountRepository: accountRepository
-//        )
-
-        context = ActionContext(state: .validateServer)
-        actions = Self.makeActions(container: container, config: config)
-
-        self.metrics = container.resolve(SDKMetrics.self)
-        self.logger = container.resolve(Logger.self)
-        self.latestBlocksDataProvider = container.resolve(LatestBlocksDataProvider.self)
-        self.internalSyncProgress = container.resolve(InternalSyncProgress.self)
-        self.blockDownloaderService = container.resolve(BlockDownloaderService.self)
-        self.service = container.resolve(LightWalletService.self)
-        self.rustBackend = container.resolve(ZcashRustBackendWelding.self)
-        self.storage = container.resolve(CompactBlockRepository.self)
-        self.config = config
-        self.transactionRepository = container.resolve(TransactionRepository.self)
-        self.accountRepository = accountRepository
-    }
-
-    // swiftlint:disable:next cyclomatic_complexity
-    private static func makeActions(container: DIContainer, config: Configuration) -> [CBPState: Action] {
-        let actionsDefinition = CBPState.allCases.compactMap { state -> (CBPState, Action)? in
-            let action: Action
-            switch state {
-            case .migrateLegacyCacheDB:
-                action = MigrateLegacyCacheDBAction(container: container, config: config)
-            case .validateServer:
-                action = ValidateServerAction(container: container, config: config)
-            case .computeSyncRanges:
-                action = ComputeSyncRangesAction(container: container, config: config)
-            case .checksBeforeSync:
-                action = ChecksBeforeSyncAction(container: container)
-            case .scanDownloaded:
-                action = ScanDownloadedButUnscannedAction(container: container)
-            case .download:
-                action = DownloadAction(container: container, config: config)
-            case .validate:
-                action = ValidateAction(container: container)
-            case .scan:
-                action = ScanAction(container: container, config: config)
-            case .clearAlreadyScannedBlocks:
-                action = ClearAlreadyScannedBlocksAction(container: container)
-            case .enhance:
-                action = EnhanceAction(container: container, config: config)
-            case .fetchUTXO:
-                action = FetchUTXOsAction(container: container)
-            case .handleSaplingParams:
-                action = SaplingParamsAction(container: container)
-            case .clearCache:
-                action = ClearCacheAction(container: container)
-            case .finished, .failed, .stopped, .idle:
-                return nil
-            }
-
-            return (state, action)
-        }
-
-        return Dictionary(uniqueKeysWithValues: actionsDefinition)
-    }
-}
-
-// MARK: - "Public" API
-
-extension CompactBlockProcessorNG {
-    func start(retry: Bool = false) async {
-        if retry {
-            self.retryAttempts = 0
-            self.backoffTimer?.invalidate()
-            self.backoffTimer = nil
-        }
-
-        guard await canStartSync() else {
-            if await isIdle() {
-                logger.warn("max retry attempts reached on \(await context.state) state")
-                await send(event: .failed(ZcashError.compactBlockProcessorMaxAttemptsReached(config.retries)))
-            } else {
-                logger.debug("Warning: compact block processor was started while busy!!!!")
-                afterSyncHooksManager.insert(hook: .anotherSync)
-            }
-            return
-        }
-
-        syncTask = Task(priority: .userInitiated) {
-            await run()
-        }
-    }
-
-    func stop() async {
-        await rawStop()
-        retryAttempts = 0
-    }
-}
-
-// MARK: - Events
-
-extension CompactBlockProcessorNG {
-    typealias EventClosure = (Event) async -> Void
-
-    enum Event {
-        /// Event sent when the CompactBlockProcessor presented an error.
-        case failed (Error)
-
-        /// Event sent when the CompactBlockProcessor has finished syncing the blockchain to latest height
-        case finished (_ lastScannedHeight: BlockHeight)
-
-        /// Event sent when the CompactBlockProcessor enhanced a bunch of transactions in some range.
-        case foundTransactions ([ZcashTransaction.Overview], CompactBlockRange)
-
-        /// Event sent when the CompactBlockProcessor handled a ReOrg.
-        /// `reorgHeight` is the height on which the reorg was detected.
-        /// `rewindHeight` is the height that the processor backed to in order to solve the Reorg.
-        case handledReorg (_ reorgHeight: BlockHeight, _ rewindHeight: BlockHeight)
-
-        /// Event sent when progress of the sync process changes.
-        case progressUpdated (CompactBlockProgress)
-
-        /// Event sent when the CompactBlockProcessor fetched utxos from lightwalletd attempted to store them.
-        case storedUTXOs ((inserted: [UnspentTransactionOutputEntity], skipped: [UnspentTransactionOutputEntity]))
-
-        /// Event sent when the CompactBlockProcessor starts enhancing of the transactions.
-        case startedEnhancing
-
-        /// Event sent when the CompactBlockProcessor starts fetching of the UTXOs.
-        case startedFetching
-
-        /// Event sent when the CompactBlockProcessor starts syncing.
-        case startedSyncing
-
-        /// Event sent when the CompactBlockProcessor stops syncing.
-        case stopped
-    }
-
-    func updateEventClosure(identifier: String, closure: @escaping (Event) async -> Void) async {
-        eventClosures[identifier] = closure
-    }
-
-    private func send(event: Event) async {
-        for item in eventClosures {
-            await item.value(event)
-        }
-    }
-}
-
-// MARK: - Main loop
-
-extension CompactBlockProcessorNG {
-    // This is main loop of the sync process. It simply takes state and try to find action which handles it. If action is found it executes the
-    // action. If action is not found then loop finishes. Thanks to this it's super easy to identify start point of sync process and end points
-    // of sync process without any side effects.
-    private func run() async {
-        resetContext()
-
-        while true {
-            // Sync is starting when the state is `idle`.
-            if await context.state == .idle {
-                await syncStarted()
-
-                if backoffTimer == nil {
-                    await setTimer()
-                }
-            }
-
-            // Try to find action for state.
-            guard let action = actions[await context.state] else {
-                if await syncFinished() {
-                    resetContext()
-                    continue
-                } else {
-                    break
-                }
-            }
-
-            do {
-                try Task.checkCancellation()
-
-                // Execute action.
-                context = try await action.run(with: context) { [weak self] event in
-                    await self?.send(event: event)
-                }
-
-                await didFinishAction()
-            } catch {
-                if Task.isCancelled {
-                    if await syncTaskWasCancelled() {
-                        // Start sync all over again
-                        resetContext()
-                    } else {
-                        // end the sync loop
-                        break
-                    }
-                } else {
-                    if await handleSyncFailure(action: action, error: error) {
-                        // Start sync all over again
-                        resetContext()
-                    } else {
-                        // end the sync loop
-                        break
-                    }
-                }
-            }
-        }
-    }
-
-    private func syncTaskWasCancelled() async -> Bool {
-        logger.info("Sync cancelled.")
-        syncTask = nil
-        await context.update(state: .stopped)
-        await send(event: .stopped)
-        return await handleAfterSyncHooks()
-    }
-
-    private func handleSyncFailure(action: Action, error: Error) async -> Bool {
-        if action.removeBlocksCacheWhenFailed {
-            await ifTaskIsNotCanceledClearCompactBlockCache()
-        }
-
-        if case let ZcashError.rustValidateCombinedChainInvalidChain(height) = error {
-            logger.error("Sync failed because of validation error: \(error)")
-            do {
-                try await validationFailed(at: BlockHeight(height))
-                // Start sync all over again
-                return true
-            } catch {
-                await failure(error)
-                return false
-            }
-        } else {
-            logger.error("Sync failed with error: \(error)")
-            await failure(error)
-            return false
-        }
-    }
-
-    // swiftlint:disable:next cyclomatic_complexity
-    private func didFinishAction() async {
-        // This is evalution of the state setup by previous action.
-        switch await context.state {
-        case .idle:
-            break
-        case .migrateLegacyCacheDB:
-            break
-        case .validateServer:
-            break
-        case .computeSyncRanges:
-            break
-        case .checksBeforeSync:
-            break
-        case .scanDownloaded:
-            break
-        case .download:
-            break
-        case .validate:
-            break
-        case .scan:
-            break
-        case .clearAlreadyScannedBlocks:
-            break
-        case .enhance:
-            await send(event: .startedEnhancing)
-        case .fetchUTXO:
-            await send(event: .startedFetching)
-        case .handleSaplingParams:
-            break
-        case .clearCache:
-            break
-        case .finished:
-            break
-        case .failed:
-            break
-        case .stopped:
-            break
-        }
-    }
-
-    private func resetContext() {
-        context = ActionContext(state: .idle)
-    }
-
-    private func syncStarted() async {
-        // handle start of the sync process
-        await send(event: .startedSyncing)
-    }
-
-    private func syncFinished() async -> Bool {
-        let newerBlocksWereMinedDuringSync = await context.syncRanges.latestBlockHeight < latestBlocksDataProvider.latestBlockHeight
-
-        retryAttempts = 0
-        consecutiveChainValidationErrors = 0
-
-        let lastScannedHeight = await latestBlocksDataProvider.latestScannedHeight
-        await send(event: .finished(lastScannedHeight))
-        await context.update(state: .finished)
-
-        // If new blocks were mined during previous sync run the sync process again
-        if newerBlocksWereMinedDuringSync {
-            return true
-        } else {
-            await setTimer()
-            return false
-        }
-    }
-
-    private func update(progress: CompactBlockProgress) async {
-        await send(event: .progressUpdated(progress))
-    }
-
-    private func validationFailed(at height: BlockHeight) async throws {
-        // cancel all Tasks
-        await rawStop()
-
-        // rewind
-        let rewindHeight = determineLowerBound(
-            errorHeight: height,
-            consecutiveErrors: consecutiveChainValidationErrors,
-            walletBirthday: config.walletBirthday
-        )
-
-        consecutiveChainValidationErrors += 1
-
-        try await rustBackend.rewindToHeight(height: Int32(rewindHeight))
-
-        try await blockDownloaderService.rewind(to: rewindHeight)
-        await internalSyncProgress.rewind(to: rewindHeight)
-
-        await send(event: .handledReorg(height, rewindHeight))
-    }
-
-    private func failure(_ error: Error) async {
-        await context.update(state: .failed)
-
-        logger.error("Fail with error: \(error)")
-        await rawStop()
-
-        self.retryAttempts += 1
-        await send(event: .failed(error))
-
-        // don't set a new timer if there are no more attempts.
-        if hasRetryAttempt() {
-            await self.setTimer()
-        }
-    }
-
-    private func handleAfterSyncHooks() async -> Bool {
-        let afterSyncHooksManager = self.afterSyncHooksManager
-        self.afterSyncHooksManager = AfterSyncHooksManager()
-
-        if let wipeContext = afterSyncHooksManager.shouldExecuteWipeHook() {
-            await doWipe(context: wipeContext)
-            return false
-        } else if let rewindContext = afterSyncHooksManager.shouldExecuteRewindHook() {
-            await doRewind(context: rewindContext)
-            return false
-        } else if afterSyncHooksManager.shouldExecuteAnotherSyncHook() {
-            logger.debug("Starting new sync.")
-            return true
-        } else {
-            return false
-        }
-    }
-}
-
-// MARK: - Rewind
-extension CompactBlockProcessorNG {
-    /// Rewinds to provided height.
-    /// - Parameter height: height to rewind to. If nil is provided, it will rescan to nearest height (quick rescan)
-    ///
-    /// - Note: If this is called while sync is in progress then the sync process is stopped first and then rewind is executed.
-    func rewind(context: AfterSyncHooksManager.RewindContext) async {
-        logger.debug("Starting rewind")
-        if await isIdle() {
-            logger.debug("Sync doesn't run. Executing rewind.")
-            await doRewind(context: context)
-        } else {
-            logger.debug("Stopping sync because of rewind")
-            afterSyncHooksManager.insert(hook: .rewind(context))
-            await stop()
-        }
-    }
-
-    private func doRewind(context: AfterSyncHooksManager.RewindContext) async {
-        logger.debug("Executing rewind.")
-        let lastDownloaded = await internalSyncProgress.latestDownloadedBlockHeight
-        let height = Int32(context.height ?? lastDownloaded)
-
-        let nearestHeight: Int32
-        do {
-            nearestHeight = try await rustBackend.getNearestRewindHeight(height: height)
-        } catch {
-            await failure(error)
-            return await context.completion(.failure(error))
-        }
-
-        // FIXME: [#719] this should be done on the rust layer, https://github.com/zcash/ZcashLightClientKit/issues/719
-        let rewindHeight = max(Int32(nearestHeight - 1), Int32(config.walletBirthday))
-
-        do {
-            try await rustBackend.rewindToHeight(height: rewindHeight)
-        } catch {
-            await failure(error)
-            return await context.completion(.failure(error))
-        }
-
-        // clear cache
-        let rewindBlockHeight = BlockHeight(rewindHeight)
-        do {
-            try await blockDownloaderService.rewind(to: rewindBlockHeight)
-        } catch {
-            return await context.completion(.failure(error))
-        }
-
-        await internalSyncProgress.rewind(to: rewindBlockHeight)
-
-        await context.completion(.success(rewindBlockHeight))
-    }
-}
-
-// MARK: - Wipe
-
-extension CompactBlockProcessorNG {
-    func wipe(context: AfterSyncHooksManager.WipeContext) async {
-        logger.debug("Starting wipe")
-        if await isIdle() {
-            logger.debug("Sync doesn't run. Executing wipe.")
-            await doWipe(context: context)
-        } else {
-            logger.debug("Stopping sync because of wipe")
-            afterSyncHooksManager.insert(hook: .wipe(context))
-            await stop()
-        }
-    }
-
-    private func doWipe(context: AfterSyncHooksManager.WipeContext) async {
-        logger.debug("Executing wipe.")
-        context.prewipe()
-
-        do {
-            try await self.storage.clear()
-            await internalSyncProgress.rewind(to: 0)
-
-            wipeLegacyCacheDbIfNeeded()
-
-            let fileManager = FileManager.default
-            if fileManager.fileExists(atPath: config.dataDb.path) {
-                try fileManager.removeItem(at: config.dataDb)
-            }
-
-            await context.completion(nil)
-        } catch {
-            await context.completion(error)
-        }
-    }
-
-    private func wipeLegacyCacheDbIfNeeded() {
-        guard let cacheDbURL = config.cacheDbURL else { return }
-        guard FileManager.default.isDeletableFile(atPath: cacheDbURL.pathExtension) else { return }
-        try? FileManager.default.removeItem(at: cacheDbURL)
-    }
-}
-
-// MARK: - Utils
-
-extension CompactBlockProcessorNG {
-    private func setTimer() async {
-        let interval = config.blockPollInterval
-        self.backoffTimer?.invalidate()
-        let timer = Timer(
-            timeInterval: interval,
-            repeats: true,
-            block: { [weak self] _ in
-                Task { [weak self] in
-                    guard let self else { return }
-                    if await self.isIdle() {
-                        if await self.canStartSync() {
-                            self.logger.debug(
-                                """
-                                Timer triggered: Starting compact Block processor!.
-                                Processor State: \(await self.context.state)
-                                latestHeight: \(try await self.transactionRepository.lastScannedHeight())
-                                attempts: \(await self.retryAttempts)
-                                """
-                            )
-                            await self.start()
-                        } else if await hasRetryAttempt() {
-                            await self.failure(ZcashError.compactBlockProcessorMaxAttemptsReached(self.config.retries))
-                        }
-                    } else {
-                        await self.latestBlocksDataProvider.updateBlockData()
-                    }
-                }
-            }
-        )
-        RunLoop.main.add(timer, forMode: .default)
-
-        self.backoffTimer = timer
-    }
-
-    private func isIdle() async -> Bool {
-        switch await context.state {
-        case .stopped, .failed, .finished, .idle:
-            return true
-        case .computeSyncRanges,
-             .checksBeforeSync,
-             .download,
-             .validate,
-             .scan,
-             .enhance,
-             .fetchUTXO,
-             .handleSaplingParams,
-             .clearCache,
-             .scanDownloaded,
-             .clearAlreadyScannedBlocks,
-             .validateServer,
-             .migrateLegacyCacheDB:
-            return false
-        }
-    }
-
-    private func canStartSync() async -> Bool {
-        return await isIdle() && hasRetryAttempt()
-    }
-
-    private func hasRetryAttempt() -> Bool {
-        retryAttempts < config.retries
-    }
-
-    private func determineLowerBound(errorHeight: Int, consecutiveErrors: Int, walletBirthday: BlockHeight) -> BlockHeight {
-        let offset = min(ZcashSDK.maxReorgSize, ZcashSDK.defaultRewindDistance * (consecutiveErrors + 1))
-        return max(errorHeight - offset, walletBirthday - ZcashSDK.maxReorgSize)
-    }
-
-    private func rawStop() async {
-        syncTask?.cancel()
-        self.backoffTimer?.invalidate()
-        self.backoffTimer = nil
-        await stopAllActions()
-    }
-
-    private func stopAllActions() async {
-        for action in actions.values {
-            await action.stop()
-        }
-    }
-
-    private func ifTaskIsNotCanceledClearCompactBlockCache() async {
-        guard !Task.isCancelled else { return }
-        let lastScannedHeight = await latestBlocksDataProvider.latestScannedHeight
-        do {
-            // Blocks download work in parallel with scanning. So imagine this scenario:
-            //
-            // Scanning is done until height 10300. Blocks are downloaded until height 10400.
-            // And now validation fails and this method is called. And `.latestDownloadedBlockHeight` in `internalSyncProgress` is set to 10400. And
-            // all the downloaded blocks are removed here.
-            //
-            // If this line doesn't happen then when sync starts next time it thinks that all the blocks are downloaded until 10400. But all were
-            // removed. So blocks between 10300 and 10400 wouldn't ever be scanned.
-            //
-            // Scanning is done until 10300 so the SDK can be sure that blocks with height below 10300 are not required. So it makes sense to set
-            // `.latestDownloadedBlockHeight` to `lastScannedHeight`. And sync will work fine in next run.
-            await internalSyncProgress.set(lastScannedHeight, .latestDownloadedBlockHeight)
-            try await clearCompactBlockCache()
-        } catch {
-            logger.error("`clearCompactBlockCache` failed after error: \(error)")
-        }
-    }
-
-    private func clearCompactBlockCache() async throws {
-        await stopAllActions()
-        try await storage.clear()
-        logger.info("Cache removed")
-    }
-}
