+++ Tests/TestUtils/Tests+Utils.swift
@@ -152,7 +152,6 @@ extension ZcashRustBackend {
     }
 }
 
-
 extension Zatoshi: CustomDebugStringConvertible {
     public var debugDescription: String {
         "Zatoshi(\(self.amount))"
